+++ ThermodynamicApi.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33110.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GasApi", "GasApi.csproj", "{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dynamicheating", "..\vsmodtemplate-1.5.7\mods\dynamicheat\dynamicheating.csproj", "{BA28E816-C9D9-429A-A064-10BF89885C8F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VintagestoryAPI", "..\vsapi\VintagestoryAPI.csproj", "{A3582245-A1D1-4135-BA90-8EFD7830DA83}"
@@ -12,6 +10,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSEssentialsMod", "..\vsessentialsmod\VSEssentialsMod.csproj", "{3730FF53-02AB-4CCD-80AB-6199DF67383B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSSurvivalMod", "..\vssurvivalmod\VSSurvivalMod.csproj", "{7D22278B-7FFC-403A-92D0-FD87C7609912}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ThermodynamicAPI", "ThermodynamicAPI.csproj", "{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,16 +22,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug + Flatworld|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug + Flatworld|Any CPU.Build.0 = Debug|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.FastBuild|Any CPU.ActiveCfg = Release|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.FastBuild|Any CPU.Build.0 = Release|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.PerfTest|Any CPU.ActiveCfg = Release|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.PerfTest|Any CPU.Build.0 = Release|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BA28E816-C9D9-429A-A064-10BF89885C8F}.Debug + Flatworld|Any CPU.ActiveCfg = Debug|Any CPU
 		{BA28E816-C9D9-429A-A064-10BF89885C8F}.Debug + Flatworld|Any CPU.Build.0 = Debug|Any CPU
 		{BA28E816-C9D9-429A-A064-10BF89885C8F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -72,6 +62,16 @@ Global
 		{7D22278B-7FFC-403A-92D0-FD87C7609912}.PerfTest|Any CPU.Build.0 = PerfTest|Any CPU
 		{7D22278B-7FFC-403A-92D0-FD87C7609912}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7D22278B-7FFC-403A-92D0-FD87C7609912}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug + Flatworld|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug + Flatworld|Any CPU.Build.0 = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.FastBuild|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.FastBuild|Any CPU.Build.0 = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.PerfTest|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.PerfTest|Any CPU.Build.0 = Debug|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
