+++ assets/thermodynamicapi/patches/acid.json
