+++ assets/thermodynamicapi/patches/burning.json
