+++ assets/thermodynamicapi/patches/coal.json
