+++ assets/thermodynamicapi/patches/door.json
