+++ assets/thermodynamicapi/patches/plants.json
