+++ src/Blocks/BlockLiquid.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GasApi.Blocks
+namespace ThermodynamicApi.Blocks
 {
     internal class BlockLiquid
     {
