+++ src/For Other Mods/GasHelper.cs
@@ -8,7 +8,7 @@ using Vintagestory.API.Util;
 using Vintagestory.API.Server;
 using Vintagestory.API.Config;
 
-namespace GasApi
+namespace ThermodynamicApi.ApiHelper
 {
     public class GasHelper : ModSystem
     {
