+++ src/System Control/EnumGasThreadPass.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GasAp
+﻿
+namespace ThermodynamicApi
 {
-    public enum EnumGasThreadPass
+    public enum EnumThermoThreadPass
     {
         TEMPERATURE,
         GAS,
